+minor fix
git-svn-id: https://svn.code.sf.net/p/remotefs/code/trunk@1081 7764f161-3e89-4957-9ae6-8e1f4830ae4a

+++ src/sendrecv.h
@@ -141,12 +141,12 @@ static inline ssize_t rfs_send_data(struct sendrecv_info *info, const void *data
 	send_token_t token = { 1, {{ 0 }} };
 	token.iov[0].iov_base = (void *)data;
 	token.iov[0].iov_len = data_len;
-	return (do_send(info, &token) == data_len ? 0 : -1);
+	return ((size_t)(do_send(info, &token)) == data_len ? 0 : -1);
 }
 
 static inline ssize_t rfs_receive_data(struct sendrecv_info *info, void *data, const size_t data_len)
 {
-	return (rfs_recv(info, (char *)data, data_len, 0) == data_len ? 0 : -1);
+	return ((size_t)(rfs_recv(info, (char *)data, data_len, 0)) == data_len ? 0 : -1);
 }
 
 ssize_t rfs_ignore_incoming_data(struct sendrecv_info *info, const size_t data_len);
